Little example to show default spring manage of exceptions

+++ spring-boot-first-web-application/src/main/java/com/in28minutes/springboot/web/springbootfirstwebapplication/controller/TodoController.java
@@ -57,6 +57,11 @@ public class TodoController {
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam("id") int id) {
+		
+		if (id ==1) {
+			throw new RuntimeException("Something went wrong");
+		}
+		
 		todoService.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
